Temporary commit
Plan is to use matrix A as sparse matrix, for space efficiency, because it is unacceptable.
Because of the difficult work with sparse matrices, we will use it just for checking the correction of the result.
In computation we will use classical iterative approach, as before the change.
Checking should be done after the computation, so the time taken for it
won't be summed in execution time.
This is done for iterative methods.
Multigrid method should be transformed in this style, and Fourier method
is still being analysed.