Provided dense matrix for Multigrid method
It seems that sparse matrix can't be used in calculations of smooth iterations.
And we need those calculations because we can't find dimension size so
easy when whole space size becomes not third root of some number powered
by 2. This method is one dim method adapted for three dim problem.
It divides the array on 2, and we should divide it by 8, which is not as
simple as just to change number. Maybe we should use 1D v-cycle to create
3D vcycle from outside.